+++ Assets/Scripts/Core.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 43f086ca47e9ce242824af20184176a7
+folderAsset: yes
+timeCreated: 1502671110
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
