+++ Assets/Scripts/Gameplay.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 565833b555df3874399ed9b220424456
+folderAsset: yes
+timeCreated: 1502671143
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
